Version 1.0.0/backend/requirements.txt
Flask
openai
psycopg2-binary
python-dotenv
gunicorn